Switch to shaders-based Heatmap and remove old worker code
