+++ .env
@@ -18,9 +18,6 @@ REACT_APP_HSDS_FALLBACK_DOMAIN="/home/reader/water"
 REACT_APP_JLAB_URL=
 REACT_APP_JLAB_FALLBACK_DOMAIN="water_224.h5"
 
-# Development-only feature flags
-REACT_APP_DEV_SHADERS=false
-
 # Rendering performance profiling (local development only)
 REACT_APP_PROFILING_ENABLED=false
