+++ .gitignore
@@ -16,4 +16,3 @@ npm-debug.log*
 *.log
 Thumbs.db
 .DS_Store
-.eslintcache
