+++ package.json
@@ -35,7 +35,6 @@
     "@visx/shape": "^1.4.0",
     "@visx/tooltip": "^1.3.0",
     "axios": "^0.21.1",
-    "comlink": "^4.3.0",
     "d3-array": "^2.11.0",
     "d3-color": "^2.0.0",
     "d3-format": "^2.0.0",
@@ -58,7 +57,6 @@
     "react-suspense-fetch": "^0.3.0",
     "react-three-fiber": "^4.2.21",
     "react-use": "^17.1.0",
-    "react-use-comlink": "^2.0.1",
     "react-window": "^1.8.6",
     "three": "^0.125.2",
     "zustand": "^3.3.1"
@@ -110,17 +108,13 @@
     "react-app-polyfill": "^2.0.0",
     "react-is": "^16.13.1",
     "react-scripts": "4.0.2",
-    "typescript": "^4.1.3",
-    "worker-loader": "^3.0.7"
+    "typescript": "^4.1.3"
   },
   "jest": {
     "restoreMocks": true,
     "transformIgnorePatterns": [
-      "/node_modules/(?!lodash-es|react-use-comlink)/.*"
-    ],
-    "moduleNameMapper": {
-      "worker-loader*": "<rootDir>/src/moduleMock.js"
-    }
+      "/node_modules/(?!lodash-es)/.*"
+    ]
   },
   "browserslist": {
     "production": [
