+++ packages/package.json
@@ -30,8 +30,7 @@
     "typescript": "^4.0.5",
     "url-loader": "^4.1.0",
     "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12",
-    "worker-loader": "^3.0.2"
+    "webpack-cli": "^3.3.12"
   },
   "browserslist": {
     "production": [
