+++ src/h5web/vis-packs/core/heatmap/HeatmapVis.module.css
@@ -4,29 +4,3 @@
   min-width: 0;
   min-height: 0;
 }
-
-.textureLoader {
-  background-color: rgba(0, 0, 0, 0.4);
-  background-image: linear-gradient(
-    to right,
-    transparent,
-    rgba(255, 255, 255, 0.5) 60%,
-    transparent
-  );
-  background-position-x: -50%;
-  background-size: 35% 100%; /* allows animating position in percent */
-  background-repeat: no-repeat;
-  transition: opacity 0.2s ease-out;
-  opacity: 0;
-}
-
-.textureLoader[data-visible] {
-  opacity: 1;
-  animation: shiftBg 1s ease-in-out infinite;
-}
-
-@keyframes shiftBg {
-  to {
-    background-position-x: 150%;
-  }
-}
