+++ src/h5web/vis-packs/core/heatmap/Mesh.tsx
@@ -1,10 +1,15 @@
-import { useMemo, ReactElement } from 'react';
+import { rgb } from 'd3-color';
+import { ReactElement, memo, useMemo } from 'react';
 import { useThree } from 'react-three-fiber';
-import { RGBFormat, MeshBasicMaterial, DataTexture } from 'three';
-import Html from '../shared/Html';
-import { useTextureData } from './hooks';
-import styles from './HeatmapVis.module.css';
-import type { Domain, ScaleType } from '../models';
+import {
+  DataTexture,
+  FloatType,
+  RedFormat,
+  RGBFormat,
+  UnsignedByteType,
+} from 'three';
+import { Domain, ScaleType } from '../models';
+import { INTERPOLATORS } from './interpolators';
 import type { ColorMap } from './models';
 
 interface Props {
@@ -12,50 +17,113 @@ interface Props {
   cols: number;
   values: number[];
   domain: Domain;
-  scaleType: ScaleType;
   colorMap: ColorMap;
-  showLoader: boolean;
+  scaleType: ScaleType;
 }
 
+const SCALE_FUNC: Record<ScaleType, (val: number) => number> = {
+  [ScaleType.Linear]: (val) => val,
+  [ScaleType.Log]: Math.log10,
+  [ScaleType.SymLog]: (val) => {
+    return Math.sign(val) * Math.log10(1 + Math.abs(val));
+  },
+};
+
+const CMAP_SIZE = 256;
+const CMAP_NORM: Record<ScaleType, number> = {
+  [ScaleType.Linear]: 0,
+  [ScaleType.Log]: 1,
+  [ScaleType.SymLog]: 2,
+};
+
 function Mesh(props: Props): ReactElement {
-  const { rows, cols, values, domain, scaleType, colorMap, showLoader } = props;
+  const { rows, cols, values, domain, colorMap, scaleType } = props;
+
+  const scaledDomain = domain.map(SCALE_FUNC[scaleType]);
+
+  const dataTexture = useMemo(() => {
+    const valuesArr = Float32Array.from(values);
+    return new DataTexture(valuesArr, cols, rows, RedFormat, FloatType);
+  }, [cols, rows, values]);
+
+  const colorMapTexture = useMemo(() => {
+    const interpolator = INTERPOLATORS[colorMap];
+
+    const colors = Uint8Array.from(
+      Array.from({ length: CMAP_SIZE }).flatMap((_, i) => {
+        const { r, g, b } = rgb(interpolator(i / (CMAP_SIZE - 1)));
+        return [r, g, b];
+      })
+    );
+
+    return new DataTexture(colors, CMAP_SIZE, 1, RGBFormat, UnsignedByteType);
+  }, [colorMap]);
+
+  const shader = {
+    uniforms: {
+      data: { value: dataTexture },
+      colorMap: { value: colorMapTexture },
+      scaleType: { value: CMAP_NORM[scaleType] },
+      min: { value: scaledDomain[0] },
+      oneOverRange: { value: 1 / (scaledDomain[1] - scaledDomain[0]) },
+    },
+    vertexShader: `
+      varying vec2 coords;
+
+      void main() {
+        coords = uv;
+        gl_Position = projectionMatrix * modelViewMatrix * vec4( position, 1.0 );
+      }
+    `,
+    fragmentShader: `
+      uniform sampler2D data;
+      uniform sampler2D colorMap;
+      uniform int scaleType;
+      uniform float min;
+      uniform float oneOverRange;
+
+      const float oneOverLog10 = 0.43429448190325176;
+      const vec4 nanColor = vec4(255, 255, 255, 1);
+
+      varying vec2 coords;
+
+      float symlog(float x) {
+        return sign(x) * log(1. + abs(x)) * oneOverLog10;
+      }
+
+      float scale(float value) {
+        if (scaleType == 1) {
+          return oneOverRange * (log(value) * oneOverLog10 - min); // Log
+        }
+
+        if (scaleType == 2) {
+          return oneOverRange * (symlog(value) - min); // SymLog
+        }
+
+        return oneOverRange * (value - min); // Linear
+      }
+
+      void main() {
+        float value = texture(data, coords).r;
+
+        if (scaleType == 1 && value <= 0.) {
+          gl_FragColor = nanColor;
+        } else {
+          gl_FragColor = texture(colorMap, vec2(scale(value), 0.5));
+        }
+      }
+    `,
+  };
 
   const { size } = useThree();
   const { width, height } = size;
 
-  const { textureData, loading } = useTextureData(
-    rows,
-    cols,
-    values,
-    domain,
-    scaleType,
-    colorMap
-  );
-
-  const material = useMemo(() => {
-    return (
-      textureData &&
-      new MeshBasicMaterial({
-        map: new DataTexture(textureData, cols, rows, RGBFormat),
-      })
-    );
-  }, [cols, rows, textureData]);
-
   return (
-    <>
-      {material && (
-        <mesh material={material}>
-          <planeGeometry attach="geometry" args={[width, height]} />
-        </mesh>
-      )}
-      {showLoader && (
-        <Html
-          className={styles.textureLoader}
-          data-visible={loading || undefined}
-        />
-      )}
-    </>
+    <mesh>
+      <planeGeometry attach="geometry" args={[width, height]} />
+      <shaderMaterial attach="material" args={[shader]} />
+    </mesh>
   );
 }
 
-export default Mesh;
+export default memo(Mesh);
