+++ src/stories/Home.stories.mdx
@@ -78,7 +78,6 @@ The components are organised in two categories:
      domain={domain}
      colorMap="Inferno"
      scaleType={ScaleType.Log}
-     showLoader={false}
      ...
    />;
    ```
